requirements.txt
# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
mkdocs-awesome-pages-plugin>=2.9.0
pyyaml>=6.0.0

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0

# AI/ML
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
